apps/api/requirements.txt
# Cairn API — Backend Dependencies
# Only what's needed. Nothing else.

# ── Web Framework ──
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-multipart==0.0.22
starlette==1.0.0

# ── Database ──
supabase==2.28.3
postgrest==2.28.3

# ── AI / ML ──
google-generativeai>=0.3.0
numpy>=1.26.0,<2.0
transformers>=4.35.0
torch>=2.1.0
torchaudio>=2.1.0

# ── Auth & Security ──
python-jose==3.5.0
passlib==1.7.4

# ── Utilities ──
python-dotenv==1.2.2
requests>=2.31.0
httpx>=0.27.0

# ── Monitoring (optional) ──
sentry-sdk>=2.0.0
